[Fix] Separate useCases dataProvidor / users
- Rename entireDataUseCaseInteractor -> userUseCaseInteractor
