--- app/src/main/java/com/example/cleanarchitecture/domainlayer/base/GetDataRequestModel.kt
@@ -1,3 +1,0 @@
-package com.example.cleanarchitecture.domainlayer.base
-
-interface GetDataRequestModel
