--- app/src/main/java/com/example/cleanarchitecture/domainlayer/base/GetDataResponseModel.kt
@@ -1,3 +1,0 @@
-package com.example.cleanarchitecture.domainlayer.base
-
-interface GetDataResponseModel
