+++ app/src/main/java/com/example/cleanarchitecture/domainlayer/di/UseCaseModule.kt
@@ -1,8 +1,8 @@
 package com.example.cleanarchitecture.domainlayer.di
 
-import com.example.cleanarchitecture.domainlayer.usecase.animal.CatImageUseCaseInteractorImpl
-import com.example.cleanarchitecture.domainlayer.usecase.quote.QuoteUseCaseInteractorImpl
-import com.example.cleanarchitecture.domainlayer.usecase.user.EntireDataUseCaseInteractorImpl
+import com.example.cleanarchitecture.domainlayer.usecase.dataprovidor.animal.CatImageUseCaseInteractorImpl
+import com.example.cleanarchitecture.domainlayer.usecase.dataprovidor.quote.QuoteUseCaseInteractorImpl
+import com.example.cleanarchitecture.domainlayer.usecase.user.UserUseCaseInteractorImpl
 import org.koin.dsl.module
 
 val useCaseModule = module {
@@ -10,5 +10,5 @@ val useCaseModule = module {
 
     single { QuoteUseCaseInteractorImpl(get()) }
 
-    single { EntireDataUseCaseInteractorImpl(get(), get()) }
+    single { UserUseCaseInteractorImpl(get(), get()) }
 }
