--- app/src/main/java/com/example/cleanarchitecture/domainlayer/requestmodel/CatImageRequestModel.kt
@@ -1,5 +1,0 @@
-package com.example.cleanarchitecture.domainlayer.requestmodel
-
-import com.example.cleanarchitecture.domainlayer.base.GetDataRequestModel
-
-data class CatImageRequestModel(val unit: Unit) : GetDataRequestModel
