+++ app/src/main/java/com/example/cleanarchitecture/domainlayer/requestmodel/dataprovidor/CatImageRequestModel.kt
@@ -1,0 +1,5 @@
+package com.example.cleanarchitecture.domainlayer.requestmodel.dataprovidor
+
+import com.example.cleanarchitecture.domainlayer.usecase.dataprovidor.base.GetDataRequestModel
+
+data class CatImageRequestModel(val unit: Unit) : GetDataRequestModel
