--- app/src/main/java/com/example/cleanarchitecture/domainlayer/responsemodel/QuoteResponseModel.kt
@@ -1,5 +1,0 @@
-package com.example.cleanarchitecture.domainlayer.responsemodel
-
-import com.example.cleanarchitecture.domainlayer.base.GetDataResponseModel
-
-data class QuoteResponseModel(val quote: String) : GetDataResponseModel
